% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Convert Raw NOAA Data to a Clean Data Frame}
\usage{
eq_clean_data(df)
}
\arguments{
\item{df}{A data frame containing raw NOAA data.}
}
\value{
A clean data frame.
}
\description{
The function takes raw NOAA data frame and returns a clean data frame that has:
(1) A date column created by uniting the year, month, day and converting it to
the Date class; (2) LATITUDE and LONGITUDE columns converted to numeric class.
}
\details{
The first function the user should run once they obtain the NOAA data set.
}
\examples{
system.file("extdata", "signif.txt.bz2", package = "quakes") \%>\%
    readr::read_delim(delim = "\\t", guess_max = 10000) \%>\%
    eq_clean_data()

}
