% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create More Informative Pop-Ups for Geographic Map}
\usage{
eq_create_label(df)
}
\arguments{
\item{df}{A clean data frame outputted by the code{link{eq_clean_data}} function.}
}
\value{
A character vector of HTML_formatted labels.
}
\description{
The function takes the dataset and creates an HTML label that can be used as
the annotation text in the leaflet map outptted by the code{link{eq_create_label}}
function.
}
\details{
The function puts together a character string for each earthquake that
shows the location (which is cleaned inside this function by the
code{link{eq_location_clean}} function), the magnitude (EQ_PRIMARY), and the
total number of deaths (TOTAL_DEATHS). If an earthquake is missing values for
any of these, both the label and the value are skipped for that element of the tag.
}
\examples{
usa_japan \%>\%
    quakes::eq_create_label()

}
