% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Format Location Name}
\usage{
eq_location_clean(df)
}
\arguments{
\item{df}{A data frame containing raw NOAA data.}
}
\value{
A data frame with formated LOCATION_NAME column.
}
\description{
The function cleans the LOCATION_NAME column by stripping out the country name
(including the colon) and converts names to title case (as opposed to all caps).
}
\details{
This function is needed for annotating visualisations.
}
\examples{
system.file("extdata", "signif.txt.bz2", package = "quakes") \%>\%
    readr::read_delim(delim = "\\t", guess_max = 10000) \%>\%
    quakes::eq_clean_data() \%>\%
    quakes::eq_location_clean()

}
