% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\alias{GeomTimelineLabel}
\title{Ggproto Object Used By geom_timeline}
\description{
Ggproto Object Used By geom_timeline

Ggproto Object Used By geom_timeline_label
}
\keyword{datasets}
