% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usa_japan}
\alias{usa_japan}
\title{Subset of the NOAA's Significant Earthquakes Dataset}
\format{The raw data was edited with \code{eq_clean_data} function and subsetted
to keep only earthquakes that occurred between 2014-01-05 and 2018-01-05 in the
USA and Japan.}
\source{
U.S. National Oceanographic and Atmospheric Administration (NOAA)
\url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}
}
\usage{
usa_japan
}
\description{
Subset of the NOAA's Significant Earthquakes Dataset
}
\examples{
\dontrun{
 usa_japan
}
}
\keyword{datasets}
